requirements.txt
qdrant-client>=1.7.0
python-dotenv>=1.0.0
httpx>=0.20.0
